requirements.txt
faiss-cpu
sentence-transformers
streamlit
transformers
torch
python-dotenv
langchain
langchain_groq
langchain-text-splitters
langchain_huggingface
numpy
tavily-python
flask
redis
pyjwt
python-docx
PyPDF2
celery
flask_limiter
flask_wtf
bleach
psycopg2